upgrading to sdl3

+++ include/GLApp/Mouse.h
@@ -5,7 +5,7 @@
 namespace GLApp {
 
 struct Mouse {
-	Tensor::int2 ipos = {};
+	Tensor::float2 ipos = {};
 	Tensor::float2 pos = {};
 	Tensor::float2 lastPos = {};
 	Tensor::float2 deltaPos = {};
